main module of the executable needs to be Main, with PostgREST.Main build fails
