+++ src/PostgREST/Main.hs
@@ -1,31 +1,26 @@
 {-# LANGUAGE CPP #-}
 
-module PostgREST.Main where
+module Main where
 
 
-import           PostgREST.App
+import           PostgREST.App                        (handleRequest)
 import           PostgREST.Config                     (AppConfig (..),
                                                        minimumPgVersion,
                                                        prettyVersion,
                                                        readOptions)
 import           PostgREST.DbStructure
-import           PostgREST.Error                      (pgErrResponse)
-import           PostgREST.Middleware
-import           PostgREST.Types                      (DbStructure)
 
 import           Control.Monad
 import           Data.Monoid                          ((<>))
 import           Data.String.Conversions              (cs)
-import           Data.Time.Clock.POSIX                (getPOSIXTime)
+
 import qualified Hasql.Query                          as H
 import qualified Hasql.Session                        as H
-import qualified Hasql.Transaction                    as HT
 import qualified Hasql.Decoders                       as HD
 import qualified Hasql.Encoders                       as HE
 import qualified Hasql.Pool                           as P
-import           Network.Wai
 import           Network.Wai.Handler.Warp
-import           Network.Wai.Middleware.RequestLogger (logStdout)
+
 import           System.IO                            (BufferMode (..),
                                                        hSetBuffering, stderr,
                                                        stdin, stdout)
@@ -81,17 +76,4 @@ main = do
     getDbStructure (cs $ configSchema conf)
 
   let dbStructure = either (error.show) id result
-  runSettings appSettings $ postgrest conf dbStructure pool
-
-postgrest :: AppConfig -> DbStructure -> P.Pool -> Application
-postgrest conf dbStructure pool =
-  let middle = (if configQuiet conf then id else logStdout) . defaultMiddle in
-
-  middle $ \ req respond -> do
-    time <- getPOSIXTime
-    body <- strictRequestBody req
-
-    let handleReq = runWithClaims conf time (app dbStructure conf body) req
-    resp <- either pgErrResponse id <$> P.use pool
-      (HT.run handleReq HT.ReadCommitted HT.Write)
-    respond resp
+  runSettings appSettings $ handleRequest conf dbStructure pool
