+++ test/Main.hs
@@ -6,7 +6,7 @@ import SpecHelper
 import qualified Hasql.Pool as P
 
 import PostgREST.DbStructure (getDbStructure)
-import PostgREST.Main (postgrest)
+import PostgREST.App (handleRequest)
 import Data.String.Conversions (cs)
 
 import qualified Feature.AuthSpec
@@ -27,8 +27,8 @@ main = do
 
   result <- P.use pool $ getDbStructure "test"
   let dbStructure = either (error.show) id result
-      withApp = return $ postgrest testCfg dbStructure pool
-      ltdApp  = return $ postgrest testLtdRowsCfg dbStructure pool
+      withApp = return $ handleRequest testCfg dbStructure pool
+      ltdApp  = return $ handleRequest testLtdRowsCfg dbStructure pool
 
   hspec $ do
     mapM_ (beforeAll_ resetDb . before withApp) specs
